Move commands to their own module
